pyproject.toml
[tool.poetry]
name = "notebooklm-openai"
version = "0.1.0"
description = ""
authors = ["madroid <madroidmaq@gmail.com>"]
readme = "README.md"
packages = [{ include = "notebooklm", from = "src" }]
include = ["src/notebooklm/prompts/*.txt"]

[tool.poetry.dependencies]
python = "^3.11"
PyPDF2 = "^3.0.1"
tqdm = "^4.66.6"
openai = "^1.52.2"
f5-tts-mlx = "^0.1.5"

python-dotenv = "^1.0.1"
weave = "^0.51.18"
pydub = "^0.25.1"

[tool.poetry.scripts]
generate_podcast = "notebooklm.notebook_lm:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
